requirements_hermes.txt
python-telegram-bot==21.*
google-generativeai
python-dotenv
